Agrego endpoint a listado de novedades
