+++ src/routes/AppRoutes.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Home from '../components/Home/Home.jsx';
 import About from '../components/About/About.jsx';
-import { News } from '../views/public/News.jsx';
+import { News } from '../views/News/News.jsx';
 import NewDetail from '../views/NewDetail/NewDetail.jsx';
 import Contact from '../views/Contact/Contact.jsx';
